Move the cat operations.
